+++ src/models/sys-msg.js
@@ -1,0 +1,6 @@
+export const MSG_TYPE_BATCH_IMPORT = 1;
+export const MSG_TITLE_BATCH_IMPORT = '批量导入任务';
+export const SOURCE_TYPE_BATCH_IMPORT = 'batch_import';
+export const MSG_STATUS_UNREAD = 0;
+export const MSG_STATUS_READ = 2;
+export const DICT_TYPE_SYS_MSG = 'sys_msg_type';